requirements.txt
python-telegram-bot==21.1
python-dotenv==1.0.0
supabase==2.4.2
flask==2.3.3
google-genai==0.1.0